`make-new.sh` improvements
Script checks with the user if it is okay to download the latest QCOW image. QCOW image is resized to 20G by default, but this can be changed easily. The QCOW link is always acquired now, because before sometimes `chromium` would not start properly. If this happens, it is just restarted to try again.